+++ packages/analytics/analytics-chart/src/utils/commonOptions.ts
@@ -1,6 +1,8 @@
 import type { ExternalTooltipContext, KChartData, TooltipState, TooltipEntry, Dataset, ChartLegendSortFn, LegendValues, EnhancedLegendItem, TooltipInteractionMode } from '../types'
-import { formatTooltipTimestampByGranularity, formatUnit } from '../utils'
+import { formatTooltipTimestampByGranularity } from '../utils'
 import { isValid } from 'date-fns'
+import { unitFormatter } from '@kong-ui-public/analytics-utilities'
+import composables from '../composables'
 import type { Chart, Point, ScatterDataPoint } from 'chart.js'
 import type { GranularityValues } from '@kong-ui-public/analytics-utilities'
 
@@ -22,7 +24,8 @@ export const lineChartTooltipBehavior = (
 
     return
   }
-
+  const { i18n } = composables.useI18n()
+  const { formatUnit } = unitFormatter({ i18n })
   const sortFn = tooltipData.chartTooltipSortFn || ((a: TooltipEntry, b: TooltipEntry) => b.rawValue - a.rawValue)
 
   if (tooltip.body && !isTooltipInteractive(tooltipData.interactionMode)) {
@@ -61,6 +64,8 @@ export const lineChartTooltipBehavior = (
 
 export const tooltipBehavior = (tooltipData: TooltipState, context: ExternalTooltipContext) : void => {
   const { tooltip } = context
+  const { i18n } = composables.useI18n()
+  const { formatUnit } = unitFormatter({ i18n })
   if (tooltip.opacity === 0 && !isTooltipInteractive(tooltipData.interactionMode)) {
     tooltipData.showTooltip = false