+++ packages/analytics/analytics-geo-map/src/composables/useLegendScale.spec.ts
@@ -1,11 +1,12 @@
 import { describe, it, expect } from 'vitest'
-import useMetricUtils from './useMetricUtils'
+import useLegendScale from './useLegendScale'
 import { ref } from 'vue'
-import type { ExploreAggregations } from '@kong-ui-public/analytics-utilities'
+import { COUNTRIES, type ExploreAggregations } from '@kong-ui-public/analytics-utilities'
 import { KUI_COLOR_BACKGROUND_NEUTRAL_WEAKER } from '@kong/design-tokens'
-import { colors } from './useMetricUtils'
+import { colors } from './useLegendScale'
+import type { MetricUnits } from '../types'
 
-describe('useMetricUtils', () => {
+describe('useLegendScale', () => {
   it.each([
     [[10], 1],
     [[10, 20], 2],
@@ -17,13 +18,16 @@ describe('useMetricUtils', () => {
   ])('generates correct length scale for %# datapoints', (datapoints, expectedScaleLength) => {
     const countryMetrics = ref<Record<string, number>>({})
     datapoints.forEach((value, index) => {
-      countryMetrics.value[`C${index}`] = value
+      const countryCode = COUNTRIES[index].code
+      countryMetrics.value[countryCode] = value
     })
     const metric = ref<ExploreAggregations>('request_count')
+    const unit = ref<MetricUnits>('count')
 
-    const { scale } = useMetricUtils({
+    const { scale } = useLegendScale({
       countryMetrics,
       metric,
+      unit,
     })
     expect(scale.value.length).toBe(expectedScaleLength)
   })
@@ -33,10 +37,12 @@ describe('useMetricUtils', () => {
       US: 1000,
     })
     const metric = ref<ExploreAggregations>('request_count')
+    const unit = ref<MetricUnits>('count')
 
-    const { legendData } = useMetricUtils({
+    const { legendData } = useLegendScale({
       countryMetrics,
       metric,
+      unit,
     })
 
     const expectedLegendData = [
@@ -59,10 +65,12 @@ describe('useMetricUtils', () => {
       FR: 5000,
     })
     const metric = ref<ExploreAggregations>('request_count')
+    const unit = ref<MetricUnits>('count')
 
-    const { legendData, getColor } = useMetricUtils({
+    const { legendData, getColor } = useLegendScale({
       countryMetrics,
       metric,
+      unit,
     })
 
     const expectedLegendData = [
@@ -84,35 +92,27 @@ describe('useMetricUtils', () => {
     expect(getColor(0)).toBe(KUI_COLOR_BACKGROUND_NEUTRAL_WEAKER)
   })
 
-  it('latency metrics not truncated', () => {
+  it('ignores country codes that are not in the COUNTRIES list', () => {
     const countryMetrics = ref({
-      US: 100,
-      CA: 200,
-      MX: 300,
-    })
-    const metric = ref<ExploreAggregations>('kong_latency_average')
-
-    const { formatMetric } = useMetricUtils({
-      countryMetrics,
-      metric,
-    })
-
-    expect(formatMetric(10000)).toBe('10,000')
-  })
-
-  it('metrics truncated properly', () => {
-    const countryMetrics = ref({
-      US: 100,
-      CA: 200,
-      MX: 300,
+      US: 1000,
+      empty: 2000,
+      YY: 3000,
     })
     const metric = ref<ExploreAggregations>('request_count')
+    const unit = ref<MetricUnits>('count')
 
-    const { formatMetric } = useMetricUtils({
+    const { legendData } = useLegendScale({
       countryMetrics,
       metric,
+      unit,
     })
 
-    expect(formatMetric(10000)).toBe('10K')
+    // Will generate just one legend item since the other two country codes are invalid and will be ignored
+    expect(legendData.value).toEqual([
+      {
+        'color': '#00819d',
+        'range': '1K',
+      },
+    ])
   })
 })