+++ packages/entities/entities-plugins/src/definitions/schemas/KeyAuth.ts
@@ -13,6 +13,8 @@ export const keyAuthCredentialSchema = {
         submitWhenNull: false,
         hint: `You can optionally set your own unique key to authenticate the
                client. If missing, it will be generated for you.`,
+        inputType: 'password',
+        encrypted: true,
       },
     },
     {