+++ packages/entities/entities-redis-configurations/src/index.ts
@@ -1,6 +1,8 @@
 import RedisConfigurationForm from './components/RedisConfigurationForm.vue'
 import RedisConfigurationList from './components/RedisConfigurationList.vue'
 import RedisConfigurationConfigCard from './components/RedisConfigurationConfigCard.vue'
+import RedisConfigurationSelector from './components/RedisConfigurationSelector.vue'
+import RedisConfigurationFormModal from './components/RedisConfigurationFormModal.vue'
 import LinkedPlugins from './components/LinkedPluginList.vue'
 import DeleteWarningModal from './components/DeleteWarningModal.vue'
 
@@ -8,6 +10,8 @@ export {
   RedisConfigurationForm,
   RedisConfigurationList,
   RedisConfigurationConfigCard,
+  RedisConfigurationSelector,
+  RedisConfigurationFormModal,
   LinkedPlugins,
   DeleteWarningModal,
 }
@@ -27,7 +31,12 @@ import {
   useLinkedPluginsFetcher,
 } from './composables/useLinkedPlugins'
 
+import {
+  useRedisConfigurationSelector,
+} from './composables/useRedisConfigurationSelector'
+
 export const composables = {
   useLinkedPlugins,
   useLinkedPluginsFetcher,
+  useRedisConfigurationSelector,
 }