+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.108.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.108.0...@kong-ui-public/entities-plugins@9.108.1) (2025-10-27)
+
+
+### Bug Fixes
+
+* **entities-plugins:** mask key field in key auth mock schema used by consumer ([#2594](https://github.com/Kong/public-ui-components/issues/2594)) ([952f7ef](https://github.com/Kong/public-ui-components/commit/952f7ef35f4df1c5b74170f8fbe0f9f485ee4a92))
+
+
+
+
+
+# [9.108.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.5...@kong-ui-public/entities-plugins@9.108.0) (2025-10-27)
+
+
+### Features
+
+* **entities-*:** datakit - resources cache [KM-1859] ([#2582](https://github.com/Kong/public-ui-components/issues/2582)) ([e11650c](https://github.com/Kong/public-ui-components/commit/e11650cc621f65db51b544475143a5c5c57a6b70))
+
+
+
+
+
+## [9.107.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.4...@kong-ui-public/entities-plugins@9.107.5) (2025-10-27)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#2593](https://github.com/Kong/public-ui-components/issues/2593)) ([1283fdd](https://github.com/Kong/public-ui-components/commit/1283fdd7e9f6bccf0fbb5df160917e53d1293a03))
+
+
+
+
+
+## [9.107.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.3...@kong-ui-public/entities-plugins@9.107.4) (2025-10-24)
+
+
+### Bug Fixes
+
+* **entities-plugins:** update zod version and use safe extend ([#2590](https://github.com/Kong/public-ui-components/issues/2590)) ([e6b97bd](https://github.com/Kong/public-ui-components/commit/e6b97bd8ab4b2c8b85b205b144a1a0c291fa7ab9))
+
+
+
+
+
+## [9.107.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.2...@kong-ui-public/entities-plugins@9.107.3) (2025-10-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.107.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.1...@kong-ui-public/entities-plugins@9.107.2) (2025-10-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
 ## [9.107.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.107.0...@kong-ui-public/entities-plugins@9.107.1) (2025-10-21)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins