+++ packages/entities/entities-vaults/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.26.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.26.3...@kong-ui-public/entities-vaults@3.26.4) (2025-10-27)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#2593](https://github.com/Kong/public-ui-components/issues/2593)) ([1283fdd](https://github.com/Kong/public-ui-components/commit/1283fdd7e9f6bccf0fbb5df160917e53d1293a03))
+
+
+
+
+
+## [3.26.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.26.2...@kong-ui-public/entities-vaults@3.26.3) (2025-10-22)
+
+
+### Bug Fixes
+
+* **entities-vaults:** mask api key field in vault detail page ([#2587](https://github.com/Kong/public-ui-components/issues/2587)) ([9f3f0c1](https://github.com/Kong/public-ui-components/commit/9f3f0c10873f37c592a81af1f808ccdfc8af48ff))
+
+
+
+
+
+## [3.26.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.26.1...@kong-ui-public/entities-vaults@3.26.2) (2025-10-22)
+
+
+### Bug Fixes
+
+* **entities-vaults:** mask token and app role seceret id fields in vault detail page ([#2583](https://github.com/Kong/public-ui-components/issues/2583)) ([d95ac6f](https://github.com/Kong/public-ui-components/commit/d95ac6fd8f7b3bc86ba5861a87f56ef4816a1fc9))
+
+
+
+
+
 ## [3.26.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.26.0...@kong-ui-public/entities-vaults@3.26.1) (2025-10-21)
 
 **Note:** Version bump only for package @kong-ui-public/entities-vaults